Migrating Log

+++ Render/Objects/Texture/Texture.c
@@ -1,7 +1,7 @@
 #include "Texture.h"
 #include "v2/_interface/VectorStructuresC.h"
 
-void Texture_Alloc_3C(Texture3c_PTR p_texture, size_t p_width, size_t p_height)
+void Texture_Alloc_3C(Texture3c_PTR p_texture, uint16_t p_width, uint16_t p_height)
 {
 	p_texture->Width = p_width;
 	p_texture->Height = p_height;
@@ -19,9 +19,9 @@ size_t Texture_GetSizeInBytes_3C(const Texture3c_PTR p_texture)
 	return ((size_t)p_texture->Width * p_texture->Height) * (3 * sizeof(char));
 };
 
-size_t Texture_GetElementOffset_3C(int W, int H, int textWidth)
+size_t Texture_GetElementOffset_3C(uint16_t W, uint16_t H, uint16_t textWidth)
 {
-	return (W * sizeof(Vector3c)) + (H * textWidth * sizeof(Vector3c));
+	return (W * sizeof(Vector3c)) + ((size_t)H * textWidth * sizeof(Vector3c));
 }
 
 void Texture_CuildClipRect_3C(Texture3c_PTR p_texture, Recti_PTR out_rect)
